글 삭제 구현 완

+++ Assets/02.Scripts/ArticleManager.cs
@@ -1,19 +1,16 @@
-using MongoDB.Bson;
-using MongoDB.Driver;
 using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.IO;
-using System.Runtime.CompilerServices;
+using System.Runtime.Serialization.Formatters.Binary;
+using MongoDB.Bson;
+using MongoDB.Driver;
 using UnityEngine;
-using UnityEngine.UI;
 
-// 1. 하나만을 보장 (싱글톤 사용)
+// 1. 하나만을 보장
 // 2. 어디서든 쉽게 접근 가능
-// 3. 데이터 관리
 public class ArticleManager : MonoBehaviour
 {
-
     // 게시글 리스트
     private List<Article> _articles = new List<Article>();
     public List<Article> Articles => _articles;
@@ -22,20 +19,14 @@ public class ArticleManager : MonoBehaviour
     private IMongoCollection<Article> _articleCollection;
 
     public static ArticleManager Instance { get; private set; }
-
-
-    public UI_ArticleList Ui_ArticleList;
-
     private void Awake()
     {
-
         Instance = this;
         Init();
         FindAll();
-
     }
 
-    // 몽고 
+    // 몽고DB -> 데이터베이스 -> 콜렉션 연결
     private void Init()
     {
         // 몽고 DB로부터 article 조회
@@ -43,60 +34,28 @@ public class ArticleManager : MonoBehaviour
         string connectionString = "mongodb+srv://Subin:SoobSoob@cluster0.xjolm4a.mongodb.net/";
         MongoClient mongoClient = new MongoClient(connectionString);
         // 2. 특정 데이터베이스 연결
-        IMongoDatabase sampleDB = mongoClient.GetDatabase("metaverse");
+        IMongoDatabase db = mongoClient.GetDatabase("metaverse");
         // 3. 특정 콜렉션 연결
-        _articleCollection = sampleDB.GetCollection<Article>("articles");
-
+        _articleCollection = db.GetCollection<Article>("articles");
     }
 
     public void FindAll()
     {
         // 4. 모든 문서 읽어오기
-        // 4-1. WriteTime을 기준으로 '정렬'
-        // sort 메서드를 이용해서 도큐먼트를 정렬할 수 있다. 
-        // 매개변수로는 어떤 Key로 정렬할 것인지 알려주는 BsonDocument를 전달해주면 된다. 
+        // 4.1 WriteTime을 기준으로 '정렬'
+        // Sort 메서드를 이용해서 도큐먼트를 정렬할 수 있다.
+        // 매개변수로는 어떤 Key로 정렬할 것인지 알려주는 BsonDocument를 전달해주면 된다.
         var sort = new BsonDocument();
         sort["WriteTime"] = -1;
         // +1 -> 오름차순 정렬 -> 낮은 값에서 높은 값으로 정렬한다.
-        // -1 -> 내림차순 정렬 -> 높은 값에서 낮은 값으로 정렬한다. 
+        // -1 -> 내림차순 정렬 -> 높은 값에서 낮은 값으로 정렬한다.
         _articles = _articleCollection.Find(new BsonDocument()).Sort(sort).ToList();
-        //var dataList = _articleCollection.Find(new BsonDocument()).Sort(sort).ToList();
-        // 5. 읽어온 문서 만큼 New Article()해서 데이터 채우고 
-        /*_articles.Clear();
-        foreach (var data in dataList)
-        {
-            Article article = new Article();
-            article.Name = data["Name"].ToString();
-            article.Content = data["Content"].ToString();
-            article.Like = (int)data["Like"];
-            article.ArticleType = (ArticleType)(int)data["ArticleType"];
-
-
-            article.WriteTime = DateTime.Parse(data["WriteTime"].ToString());
-            //    _articles에 넣기
-            _articles.Add(article);
-        }
-        */
     }
 
     public void FindNotice()
     {
         // 4. 공지 문서 읽어오기
         _articles = _articleCollection.Find(data => (int)data.ArticleType == (int)ArticleType.Notice).ToList();
-        // 5. 읽어온 문서 만큼 New Article()해서 데이터 채우고 
-        /*_articles.Clear();
-        foreach (var data in dataList)
-        {
-            Article article = new Article();
-            article.Name = data["Name"].ToString();
-            article.Content = data["Content"].ToString();
-            article.Like = (int)data["Like"];
-            article.ArticleType = (ArticleType)(int)data["ArticleType"];
-            article.WriteTime = DateTime.Parse(data["WriteTime"].ToString());
-            //    _articles에 넣기
-            _articles.Add(article);
-        }
-        */
     }
 
     public void Write(ArticleType articleType, string content)
@@ -106,18 +65,31 @@ public class ArticleManager : MonoBehaviour
             ArticleType = articleType,
             Name = "정수빈",
             Content = content,
-            Like = 999,
+            Like = 0,
             WriteTime = DateTime.Now
         };
 
         _articleCollection.InsertOne(article);
     }
 
+
     public void Delete(ObjectId id)
     {
+
+        if( == null)
+        {
+            return;
+        }
         //var filter = Builders<Article>.Filter.Eq("_id", id);
         //_articleCollection.DeleteOne(filter);
         _articleCollection.DeleteOne(d => d.MyID == id);
     }
+
+
+
+
+
+
+
 
 }
